Throwaway, experimental commit: adding a kernel ID to all cells.
Kernel ID doesn't get regenerated when the kernel is reset. Neither does
the client session ID. So this won't work for making sure that slices
don't slice into previous kernel sessions. The solution for that is
probably to stop including cells after the execution count hits 1 (or,
more simply, when the execution count stops decreasing).